rtimers and xmac seems to work (at least it's doing stuff)
I had to change the rtimer type to uint32 however.

+++ core/sys/rtimer.h
@@ -53,7 +53,9 @@
 #ifndef __RTIMER_H__
 #define __RTIMER_H__
 
-typedef unsigned short rtimer_clock_t;
+#include <stdint.h>
+//typedef unsigned short rtimer_clock_t;
+typedef uint32_t rtimer_clock_t;
 #define RTIMER_CLOCK_LT(a,b)     ((signed short)((a)-(b)) < 0)
 
 /**
